Update ESLint setting and Shared Package
