+++ apps/next-admin-portal/package.json
@@ -139,6 +139,8 @@
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.2.0",
     "@types/sortablejs": "^1.15.8",
+    "@typescript-eslint/eslint-plugin": "^7.10.0",
+    "@typescript-eslint/parser": "^7.10.0",
     "@vitejs/plugin-react": "^4.3.0",
     "@vitest/coverage-v8": "^1.6.0",
     "@vitest/ui": "^1.6.0",
@@ -150,7 +152,10 @@
     "eslint-config-next": "^14.2.4",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-only-warn": "^1.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-simple-import-sort": "^12.1.0",
     "husky": "^9.0.11",
     "jsdom": "^24.1.0",
