+++ apps/next-admin-portal/src/modules/files/api/files.api.ts
@@ -1,5 +1,4 @@
 import { EntityId } from '@reduxjs/toolkit';
-import { objectToQueryString } from '~shared-client/utils/querystring.util';
 
 import {
   CreateFileDto,
@@ -12,9 +11,10 @@ import {
 
 import { API_ENDPOINTS } from '@/constants/api-endpoint.constant';
 
-import { removeUndefined } from '@/utils/object.util';
-
 import axiosClient from '@/http/http-request';
+
+import { removeUndefined } from '~shared-universal/utils/object.util';
+import { objectToQueryString } from '~shared-universal/utils/string.util';
 
 export const list = (filter: FileFilter) => {
   const url = API_ENDPOINTS.FILES + '?' + objectToQueryString(removeUndefined(filter));
