+++ apps/next-admin-portal/src/modules/products/api/products.api.ts
@@ -1,13 +1,13 @@
 import { EntityId } from '@reduxjs/toolkit';
-import { objectToQueryString } from '~shared-client/utils/querystring.util';
 
 import { ProductFilter, ProductFormData, ProductResponse, ProductsResponse } from '../interfaces/products.interface';
 
 import { API_ENDPOINTS } from '@/constants/api-endpoint.constant';
 
-import { removeUndefined } from '@/utils/object.util';
-
 import axiosClient from '@/http/http-request';
+
+import { removeUndefined } from '~shared-universal/utils/object.util';
+import { objectToQueryString } from '~shared-universal/utils/string.util';
 
 export const list = (filter: ProductFilter) => {
   const url = API_ENDPOINTS.PRODUCTS + '?' + objectToQueryString(removeUndefined(filter));
