+++ apps/next-admin-portal/src/modules/products/components/product-form-category.tsx
@@ -10,7 +10,7 @@ import { ProductFormData } from '../interfaces/products.interface';
 
 import { CategoryEntity } from '@/modules/categories/interfaces/categories.interface';
 
-import { repeatStr } from '@/utils/string.util';
+import { repeatStr } from '~shared-universal/utils/string.util';
 
 const renderCategories = (cates: CategoryEntity[], depth = 0) => {
   return cates.map(category => (
