+++ apps/next-app/src/modules/blogs/api/posts.api.ts
@@ -1,11 +1,12 @@
 import { EntityId } from '@reduxjs/toolkit';
-import { objectToQueryString } from '~shared-client/utils/querystring.util';
 
 import { PostEntity, PostFilter, PostResponse, PostsResponse } from '../interfaces/posts.interface';
 
 import { API_ENDPOINTS } from '@/constants/api-endpoint.constant';
 
 import axiosClient from '@/http/http-request';
+
+import { objectToQueryString } from '~shared-universal/utils/string.util';
 
 export const list = (filter: PostFilter) => {
   const url = [API_ENDPOINTS.POSTS];
