--- apps/next-app/src/utils/object.util.ts
@@ -1,11 +1,0 @@
-export function removeUndefined<T extends object>(obj: T): Partial<T> {
-  const result: Partial<T> = {};
-
-  for (const key in obj) {
-    if (obj[key] !== undefined) {
-      result[key] = obj[key];
-    }
-  }
-
-  return result;
-}
