+++ apps/react-native-app/metro.config.js
@@ -20,6 +20,7 @@ config.resolver.nodeModulesPaths = [
   path.resolve(workspaceRoot, 'node_modules'),
   path.resolve(projectRoot, 'node_modules'),
   path.resolve(projectRoot, 'node_modules/@repo/shared-validators'),
+  path.resolve(projectRoot, 'node_modules/@repo/shared-universal'),
   path.resolve(projectRoot, 'node_modules/@repo/react-native-design-system')
 ];
