--- apps/react-native-app/src/utils/string.util.ts
@@ -1,3 +1,0 @@
-export function toCapitalized(text: string) {
-  return text.charAt(0).toUpperCase() + text.slice(1);
-}
