+++ apps/react-native-app/tsconfig.json
@@ -8,6 +8,7 @@
       "@tests/*": ["./__tests__/*"],
       "@mocks/*": ["./__mocks__/*"],
       "~shared-validators/*": ["../../packages/shared-validators/src/*"],
+      "~shared-universal/*": ["../../packages/shared-universal/src/*"],
       "~react-native-design-system": ["../../packages/react-native-design-system/src/index.ts"],
       "~react-native-design-system/*": ["../../packages/react-native-design-system/src/*"]
     }
@@ -19,6 +20,7 @@
     "__tests__",
     "__mocks__",
     "../../packages/shared-validators/src",
+    "../../packages/shared-universal/src",
     "../../packages/react-native-design-system/src"
   ],
   "exclude": ["**/Pods/**"]
