+++ packages/react-native-design-system/package.json
@@ -13,6 +13,7 @@
   "devDependencies": {
     "@repo/config-typescript": "*",
     "@repo/eslint-config": "*",
+    "@repo/shared-universal": "*",
     "typescript": "^5.4.5"
   }
 }
