+++ packages/shared-client/package.json
@@ -14,12 +14,12 @@
     "test:uicov": "vitest --ui --coverage.enabled"
   },
   "exports": {
+    "./hooks/*": "./src/hooks/*.ts",
+    "./utils/*": "./src/utils/*.ts",
     "./components/*": [
       "./src/components/*.tsx",
       "./src/components/*.ts"
-    ],
-    "./hooks/*": "./src/hooks/*.ts",
-    "./utils/*": "./src/utils/*.ts"
+    ]
   },
   "devDependencies": {
     "@repo/config-tailwindcss": "*",
@@ -36,7 +36,10 @@
     "eslint": "^8.57.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-only-warn": "^1.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-simple-import-sort": "^12.1.0",
     "msw": "^2.3.0",
     "postcss": "^8.4.38",
@@ -44,5 +47,8 @@
     "tailwindcss": "^3.4.3",
     "typescript": "^5.4.5",
     "vitest": "^1.6.0"
+  },
+  "dependencies": {
+    "slugify": "^1.6.6"
   }
 }
