+++ packages/shared-universal/.eslintignore
@@ -1,0 +1,3 @@
+node_modules
+unit-coverage
+unit-results
