+++ packages/shared-universal/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@repo/shared-universal",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Common utilities, hooks",
+  "license": "MIT",
+  "scripts": {
+    "lint": "eslint . --max-warnings 0",
+    "lint:fix": "eslint . --fix"
+  },
+  "exports": {
+    "./utils/*": "./src/utils/*.ts",
+    "./hooks/*": "./src/hooks/*.ts"
+  },
+  "devDependencies": {
+    "@repo/config-typescript": "*",
+    "@repo/eslint-config": "*",
+    "@types/node": "^20.12.11",
+    "@types/react": "^18.3.3",
+    "eslint": "^8.57.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-only-warn": "^1.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.34.1",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-simple-import-sort": "^12.1.0",
+    "typescript": "^5.4.5"
+  },
+  "dependencies": {
+    "slugify": "^1.6.6"
+  }
+}
