+++ packages/shared-universal/src/hooks/use-deep-compare-effect.ts
@@ -1,10 +1,10 @@
 import { DependencyList, EffectCallback, useEffect, useRef } from 'react';
-import { isEqual } from 'lodash-es';
+import { isDeepEqual } from '../utils/object.util';
 
 export default function useDeepCompareEffect(callback: EffectCallback, dependencies?: DependencyList) {
   const currentDependenciesRef = useRef<DependencyList>();
 
-  if (!isEqual(currentDependenciesRef.current, dependencies)) {
+  if (!isDeepEqual(currentDependenciesRef.current, dependencies)) {
     currentDependenciesRef.current = dependencies;
   }
   useEffect(callback, [currentDependenciesRef.current]);
