+++ packages/shared-universal/src/hooks/use-error-state.ts
