+++ packages/shared-validators/.eslintrc.js
@@ -3,6 +3,7 @@ module.exports = {
   root: true,
   extends: ['@repo/eslint-config/react-internal.js'],
   parser: '@typescript-eslint/parser',
+  ignorePatterns: ['prettier.config.js', 'vitest.config.ts'],
   rules: {
     'no-redeclare': 'off'
   }
